refactor: colocate pane unit tests
